refactor: 1 file for 1 component, destructure auth context
