+++ src/App.tsx
@@ -1,47 +1,12 @@
 import React from "react";
-import { Routes, Route, Outlet, useNavigate } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
+
 import Login from "./login/Login";
 import RouteProtector from "./login/components/RouteProtector";
-import AuthProvider, { AuthContext } from "./login/components/AuthProvider";
-
-function Cashier() {
-  return (
-    <>
-      <h1>Cashier page</h1>
-    </>
-  );
-}
-
-function Manager() {
-  return (
-    <>
-      <h1>Manager page</h1>
-    </>
-  );
-}
-
-function Dashboard() {
-  const auth = React.useContext(AuthContext);
-  const navigate = useNavigate();
-
-  return (
-    <>
-      <main>
-        <h2>Welcome to the dashboard page, {auth.user?.login}!</h2>
-        <button
-          onClick={() => {
-            auth.signout(() => {
-              navigate("/");
-            });
-          }}
-        >
-          Log out
-        </button>
-        <Outlet />
-      </main>
-    </>
-  );
-}
+import AuthProvider from "./login/components/AuthProvider";
+import DashboardLayout from "./common/components/DashboardLayout";
+import Manager from "./manager/components/Manager";
+import Cashier from "./cashier/components/Cashier";
 
 function App() {
   return (
@@ -49,7 +14,7 @@ function App() {
       <AuthProvider>
         <Routes>
           <Route path="/" element={<Login />} />
-          <Route path="/dashboard" element={<Dashboard />}>
+          <Route path="/dashboard" element={<DashboardLayout />}>
             <Route
               path="cashier"
               element={
