+++ src/cashier/components/Cashier.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Cashier() {
+  return (
+    <>
+      <h1>Cashier page</h1>
+    </>
+  );
+}
+
+export default Cashier;
