+++ src/index.css
@@ -1,4 +1,6 @@
-body {
-    margin: 0;
-    padding: 0;
+*,
+*::after,
+*::before {
+  margin: 0;
+  padding: 0;
 }
