+++ src/index.tsx
@@ -1,8 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
+
+import App from "./App";
+
 import "./index.css";
-import App from "./App"; 
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
