+++ src/login/components/RouteProtector.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useLocation, Navigate } from "react-router-dom";
+
 import { AuthContext } from "./AuthProvider";
 
 function RouteProtector({
@@ -9,14 +10,14 @@ function RouteProtector({
   children: JSX.Element;
   role?: string;
 }) {
-  const auth = React.useContext(AuthContext);
+  const { user } = React.useContext(AuthContext);
   const location = useLocation();
 
-  if (!auth.user || role !== auth.user.role) {
+  if (!user || role !== user.role) {
     return <Navigate to="/" state={{ from: location }} replace />;
   }
 
   return children;
-};
+}
 
 export default RouteProtector;
