+++ src/manager/components/Manager.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Manager() {
+  return (
+    <>
+      <h1>Manager page</h1>
+    </>
+  );
+}
+
+export default Manager;
